+++ LeafLog/LeafLog/App/Flows/MyInfoTabFlow.swift
@@ -8,6 +8,7 @@
 import UIKit
 import RxFlow
 import ReactorKit
+import PhotosUI
 
 final class MyInfoTabFlow: Flow {
     private let navigationController = UINavigationController()
@@ -46,6 +47,10 @@ final class MyInfoTabFlow: Flow {
 
             navigationController.present(alert, animated: true)
             return .none
+            
+        case .profileImageSourceSheet:
+            presentProfileImageSourceSheet()
+            return .none
 
         case .loginRequired:
             return .end(forwardToParentFlowWithStep: step)
@@ -55,3 +60,55 @@ final class MyInfoTabFlow: Flow {
         }
     }
 }
+
+// MARK: - Helper
+private extension MyInfoTabFlow {
+    func presentProfileImageSourceSheet() {
+        guard let profileEditVC = navigationController.topViewController as? ProfileEditViewController else {
+            return
+        }
+
+        let alertController = UIAlertController(title: nil, message: nil, preferredStyle: .actionSheet)
+
+        if UIImagePickerController.isSourceTypeAvailable(.camera) {
+            alertController.addAction(UIAlertAction(title: "카메라", style: .default) { [weak self] _ in
+                self?.presentCameraPicker(from: profileEditVC)
+            })
+        }
+
+        alertController.addAction(UIAlertAction(title: "앨범", style: .default) { [weak self] _ in
+            self?.presentPhotoPicker(from: profileEditVC)
+        })
+
+        alertController.addAction(UIAlertAction(title: "취소", style: .cancel))
+
+        if let popoverController = alertController.popoverPresentationController {
+            popoverController.sourceView = profileEditVC.profileImagePickerSourceView
+            popoverController.sourceRect = profileEditVC.profileImagePickerSourceView.bounds
+        }
+
+        navigationController.present(alertController, animated: true)
+    }
+
+    func presentCameraPicker(from viewController: ProfileEditViewController) {
+        guard UIImagePickerController.isSourceTypeAvailable(.camera) else { return }
+
+        let picker = UIImagePickerController()
+        picker.sourceType = .camera
+        picker.delegate = viewController
+
+        navigationController.present(picker, animated: true)
+    }
+
+    func presentPhotoPicker(from viewController: ProfileEditViewController) {
+        var configuration = PHPickerConfiguration(photoLibrary: .shared())
+        configuration.filter = .images
+        configuration.selectionLimit = 1
+
+        let picker = PHPickerViewController(configuration: configuration)
+        picker.delegate = viewController
+
+        navigationController.present(picker, animated: true)
+    }
+}
+
